Drop support for long as we have BigInt.
