+++ package.json
@@ -33,7 +33,6 @@
     "benchmark": "^2.1.4",
     "compile-json-stringify": "^0.1.2",
     "is-my-json-valid": "^2.20.0",
-    "long": "^4.0.0",
     "moment": "^2.24.0",
     "pre-commit": "^1.2.2",
     "proxyquire": "^2.1.3",
