+++ test/integer.test.js
@@ -2,10 +2,8 @@
 
 const t = require('tap')
 const test = t.test
-const semver = require('semver')
 const validator = require('is-my-json-valid')
-const proxyquire = require('proxyquire')
-const build = proxyquire('..', { long: null })
+const build = require('..')
 
 test('render an integer as JSON', (t) => {
   t.plan(2)
@@ -130,10 +128,3 @@ test('render an object with an additionalProperty of type integer as JSON', (t) 
   t.equal(output, '{"num":1615}')
   t.ok(validate(JSON.parse(output)), 'valid schema')
 })
-
-if (semver.gt(process.versions.node, '10.3.0')) {
-  require('./bigint')(t.test, build)
-} else {
-  t.pass('Skip because Node version < 10.4')
-  t.end()
-}
